Tweak prompt generation script

+++ cross_domain_fsl/methods/prompt_learner.py
@@ -5,17 +5,18 @@ import torch.nn as nn
 
 _tokenizer = _Tokenizer()
 
+
 class PromptLearner(nn.Module):
     def __init__(self, classnames, clip_model):
         super().__init__()
         n_cls = len(classnames)
-        CLASS_TOKEN_POSITION = 'styler'
+        CLASS_TOKEN_POSITION = "styler"
         CSC = False
         n_ctx = 1
         ctx_init = False
         dtype = clip_model.dtype
         ctx_dim = clip_model.ln_final.weight.shape[0]
-        device = 'cuda'
+        device = "cuda"
         # clip_imsize = clip_model.visual.input_resolution
         # cfg_imsize = INPUT.SIZE[0]
         # assert cfg_imsize == clip_imsize, f"cfg_imsize ({cfg_imsize}) must equal to clip_imsize ({clip_imsize})"
@@ -49,22 +50,34 @@ class PromptLearner(nn.Module):
         classnames = [name.replace("_", " ") for name in classnames]
         name_lens = [len(_tokenizer.encode(name)) for name in classnames]
         # prompts = [prompt_prefix + " " + name + "." for name in classnames]
-        prompts_content = ['a {} style of a {}'.format(prompt_prefix, name) for name in classnames]
+        prompts_content = [
+            "a {} style of a {}".format(prompt_prefix, name) for name in classnames
+        ]
         prompts_style = [prompt_prefix]
 
-        tokenized_prompts_content = torch.cat([clip.tokenize(p) for p in prompts_content]).to(device)
+        tokenized_prompts_content = torch.cat(
+            [clip.tokenize(p) for p in prompts_content]
+        ).to(device)
         tokenized_prompts_style = torch.cat([clip.tokenize(prompts_style)]).to(device)
         with torch.no_grad():
-            embedding_contnet = clip_model.token_embedding(tokenized_prompts_content).type(dtype)
-            embedding_style = clip_model.token_embedding(tokenized_prompts_style).type(dtype)
+            embedding_contnet = clip_model.token_embedding(
+                tokenized_prompts_content
+            ).type(dtype)
+            embedding_style = clip_model.token_embedding(tokenized_prompts_style).type(
+                dtype
+            )
 
         # These token vectors will be saved when in save_model(),
         # but they should be ignored in load_model() as we want to use
         # those computed using the current class names
         self.register_buffer("token_prefix_content", embedding_contnet[:, :1, :])  # SOS
-        self.register_buffer("token_suffix_content", embedding_contnet[:, 1 + n_ctx :, :])  # CLS, EOS
+        self.register_buffer(
+            "token_suffix_content", embedding_contnet[:, 1 + n_ctx :, :]
+        )  # CLS, EOS
         self.register_buffer("token_prefix_style", embedding_style[:, :1, :])  # SOS
-        self.register_buffer("token_suffix_style", embedding_style[:, 1 + n_ctx :, :])  # CLS, EOS
+        self.register_buffer(
+            "token_suffix_style", embedding_style[:, 1 + n_ctx :, :]
+        )  # CLS, EOS
 
         self.n_cls = n_cls
         self.n_ctx = n_ctx
@@ -137,19 +150,19 @@ class PromptLearner(nn.Module):
         #     prompts = torch.cat(prompts, dim=0)
 
         if self.class_token_position == "styler":
-            style_position = 1 # not sure
+            style_position = 1  # not sure
             prompts_content = []
             for i in range(self.n_cls):
                 prefix_i = prefix_content[i : i + 1, :, :]
                 class_i = suffix_content[i : i + 1, :style_position, :]
                 suffix_i = suffix_content[i : i + 1, style_position:, :]
-                ctx_i = ctx[i : i + 1, :, :] # why expand
+                ctx_i = ctx[i : i + 1, :, :]  # why expand
                 prompt = torch.cat(
                     [
-                        prefix_i,     # (1, 1, dim)
-                        class_i,      # (1, name_len, dim)
+                        prefix_i,  # (1, 1, dim)
+                        class_i,  # (1, name_len, dim), shouldn't class be after ctx
                         ctx_i,  # (1, n_ctx, dim)
-                        suffix_i,     # (1, *, dim)
+                        suffix_i,  # (1, *, dim)
                     ],
                     dim=1,
                 )
@@ -157,34 +170,34 @@ class PromptLearner(nn.Module):
             prompts_content = torch.cat(prompts_content, dim=0)
         else:
             raise ValueError
-        
+
         prefix_i = prefix_style[0, :, :]
         suffix_i = suffix_style[0, :, :]
         ctx_i = ctx[0, :, :]
         prompts = torch.cat(
             [
-                prefix_i,     # (1, 1, dim)
+                prefix_i,  # (1, 1, dim)
                 ctx_i,  # (1, n_ctx, dim)
-                suffix_i,     # (1, *, dim)
+                suffix_i,  # (1, *, dim)
             ],
             dim=0,
         )
         prompts = prompts.unsqueeze(0)
 
         return prompts, prompts_content
-    
+
 
 class PromptLearner2(nn.Module):
     def __init__(self, classnames, clip_model):
         super().__init__()
         n_cls = len(classnames)
-        CLASS_TOKEN_POSITION = 'styler'
+        CLASS_TOKEN_POSITION = "styler"
         CSC = False
         n_ctx = 1
         ctx_init = False
         dtype = clip_model.dtype
         ctx_dim = clip_model.ln_final.weight.shape[0]
-    
+
         if ctx_init:
             # use given words to initialize context vectors
             ctx_init = ctx_init.replace("_", " ")
@@ -230,7 +243,7 @@ class PromptLearner2(nn.Module):
         self.n_ctx = n_ctx
         self.tokenized_prompts = tokenized_prompts  # torch.Tensor
         self.name_lens = name_lens
-        self.class_token_position = 'front'
+        self.class_token_position = "front"
 
     def forward(self):
         ctx = self.ctx
@@ -244,7 +257,7 @@ class PromptLearner2(nn.Module):
             prompts = torch.cat(
                 [
                     prefix,  # (n_cls, 1, dim)
-                    ctx,     # (n_cls, n_ctx, dim)
+                    ctx,  # (n_cls, n_ctx, dim)
                     suffix,  # (n_cls, *, dim)
                 ],
                 dim=1,
@@ -262,11 +275,11 @@ class PromptLearner2(nn.Module):
                 ctx_i_half2 = ctx[i : i + 1, half_n_ctx:, :]
                 prompt = torch.cat(
                     [
-                        prefix_i,     # (1, 1, dim)
+                        prefix_i,  # (1, 1, dim)
                         ctx_i_half1,  # (1, n_ctx//2, dim)
-                        class_i,      # (1, name_len, dim)
+                        class_i,  # (1, name_len, dim)
                         ctx_i_half2,  # (1, n_ctx//2, dim)
-                        suffix_i,     # (1, *, dim)
+                        suffix_i,  # (1, *, dim)
                     ],
                     dim=1,
                 )
@@ -284,8 +297,8 @@ class PromptLearner2(nn.Module):
                 prompt = torch.cat(
                     [
                         prefix_i,  # (1, 1, dim)
-                        class_i,   # (1, name_len, dim)
-                        ctx_i,     # (1, n_ctx, dim)
+                        class_i,  # (1, name_len, dim)
+                        ctx_i,  # (1, n_ctx, dim)
                         suffix_i,  # (1, *, dim)
                     ],
                     dim=1,
